requirements.txt
discord.py[voice]
bs4
youtube-search
quotes_generator
humanize
wavelink
pytesseract
Pillow
textblob
pymongo
snscrape
dnspython
PyNaCl
spotipy
pytimeparse
jishaku
unidecode
mdbdict
nltk
pytz